[Mellanox] Change drop_counters infrastructure - Change folder name to drop_packets - Add drop_packets file to contain common drops - Change test_drop_counters.py to fit the new infra - Add feature_facts.py module
