+++ tests/drop_packets/acl_templates/acl_rule_del.json
