+++ tests/drop_packets/combined_drop_counters.yml
