+++ tests/drop_packets/fanout/mellanox/exec_template.yml
