+++ tests/drop_packets/fanout/mellanox/mlnx_del_of_rule.j2
